Repo 추상화
