+++ src/main/java/antigravity/repository/CommonRepository.java
@@ -1,0 +1,8 @@
+package antigravity.repository;
+
+import java.util.Optional;
+
+public interface CommonRepository<T> {
+
+    Optional<T> findById(Long id);
+}
